movie-api/requirements.txt
pip install -r requirements.txt

# Framework principal
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de datos
psycopg[binary]==3.1.13
psycopg2-binary==2.9.9 # Necesario para el modelo ML
python-dotenv==1.0.0

# AWS
boto3==1.34.34
botocore==1.34.34

# Machine Learning y NLP
scikit-learn==1.4.0

# Google Gemini AI
google-generativeai==0.8.3

# Visualización y análisis de datos
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4
numpy==1.25.2

# Validación de datos
pydantic==2.5.3

# Utilidades HTTP
requests==2.31.0